refactor: move retryable request log write helpers out of usage_db
Keep usage_db.go under the 1200-line structure-scan limit.
